+++ StoreASP/Controllers/ClientsController.cs
@@ -27,7 +27,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Clients/Details/5
-        public async Task<IActionResult> Details(decimal? id)
+        public async Task<IActionResult> Details(long? id)
         {
             if (id == null || _context.Clients == null)
             {
@@ -67,7 +67,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Clients/Edit/5
-        public async Task<IActionResult> Edit(decimal? id)
+        public async Task<IActionResult> Edit(long? id)
         {
             if (id == null || _context.Clients == null)
             {
@@ -87,7 +87,7 @@ namespace StoreASP.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(decimal id, [Bind("IdClient,FioClient,Number,Email")] Client client)
+        public async Task<IActionResult> Edit(long id, [Bind("IdClient,FioClient,Number,Email")] Client client)
         {
             if (id != client.IdClient)
             {
@@ -118,7 +118,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Clients/Delete/5
-        public async Task<IActionResult> Delete(decimal? id)
+        public async Task<IActionResult> Delete(long? id)
         {
             if (id == null || _context.Clients == null)
             {
@@ -138,7 +138,7 @@ namespace StoreASP.Controllers
         // POST: Clients/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(decimal id)
+        public async Task<IActionResult> DeleteConfirmed(long id)
         {
             if (_context.Clients == null)
             {
@@ -154,7 +154,7 @@ namespace StoreASP.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool ClientExists(decimal id)
+        private bool ClientExists(long id)
         {
           return (_context.Clients?.Any(e => e.IdClient == id)).GetValueOrDefault();
         }
