+++ StoreASP/Controllers/MarksController.cs
@@ -27,7 +27,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Marks/Details/5
-        public async Task<IActionResult> Details(decimal? id)
+        public async Task<IActionResult> Details(long? id)
         {
             if (id == null || _context.Marks == null)
             {
@@ -67,7 +67,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Marks/Edit/5
-        public async Task<IActionResult> Edit(decimal? id)
+        public async Task<IActionResult> Edit(long? id)
         {
             if (id == null || _context.Marks == null)
             {
@@ -87,7 +87,7 @@ namespace StoreASP.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(decimal id, [Bind("IdMark,NazvanieMark")] Mark mark)
+        public async Task<IActionResult> Edit(long id, [Bind("IdMark,NazvanieMark")] Mark mark)
         {
             if (id != mark.IdMark)
             {
@@ -118,7 +118,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Marks/Delete/5
-        public async Task<IActionResult> Delete(decimal? id)
+        public async Task<IActionResult> Delete(long? id)
         {
             if (id == null || _context.Marks == null)
             {
@@ -138,7 +138,7 @@ namespace StoreASP.Controllers
         // POST: Marks/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(decimal id)
+        public async Task<IActionResult> DeleteConfirmed(long id)
         {
             if (_context.Marks == null)
             {
@@ -154,7 +154,7 @@ namespace StoreASP.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool MarkExists(decimal id)
+        private bool MarkExists(long id)
         {
           return (_context.Marks?.Any(e => e.IdMark == id)).GetValueOrDefault();
         }
