+++ StoreASP/Controllers/PostavkaProductsController.cs
@@ -26,7 +26,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: PostavkaProducts/Details/5
-        public async Task<IActionResult> Details(decimal? id)
+        public async Task<IActionResult> Details(long? id)
         {
             if (id == null || _context.PostavkaProducts == null)
             {
@@ -72,7 +72,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: PostavkaProducts/Edit/5
-        public async Task<IActionResult> Edit(decimal? id)
+        public async Task<IActionResult> Edit(long? id)
         {
             if (id == null || _context.PostavkaProducts == null)
             {
@@ -94,7 +94,7 @@ namespace StoreASP.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(decimal id, [Bind("IdPostavka,IdProduct,KolVo")] PostavkaProduct postavkaProduct)
+        public async Task<IActionResult> Edit(long id, [Bind("IdPostavka,IdProduct,KolVo")] PostavkaProduct postavkaProduct)
         {
             if (id != postavkaProduct.IdPostavka)
             {
@@ -127,7 +127,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: PostavkaProducts/Delete/5
-        public async Task<IActionResult> Delete(decimal? id)
+        public async Task<IActionResult> Delete(long? id)
         {
             if (id == null || _context.PostavkaProducts == null)
             {
@@ -149,7 +149,7 @@ namespace StoreASP.Controllers
         // POST: PostavkaProducts/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(decimal id)
+        public async Task<IActionResult> DeleteConfirmed(long id)
         {
             if (_context.PostavkaProducts == null)
             {
@@ -165,7 +165,7 @@ namespace StoreASP.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool PostavkaProductExists(decimal id)
+        private bool PostavkaProductExists(long id)
         {
           return (_context.PostavkaProducts?.Any(e => e.IdPostavka == id)).GetValueOrDefault();
         }
