+++ StoreASP/Controllers/PprovidersController.cs
@@ -27,7 +27,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Pproviders/Details/5
-        public async Task<IActionResult> Details(decimal? id)
+        public async Task<IActionResult> Details(long? id)
         {
             if (id == null || _context.Pproviders == null)
             {
@@ -67,7 +67,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Pproviders/Edit/5
-        public async Task<IActionResult> Edit(decimal? id)
+        public async Task<IActionResult> Edit(long? id)
         {
             if (id == null || _context.Pproviders == null)
             {
@@ -87,7 +87,7 @@ namespace StoreASP.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(decimal id, [Bind("IdProvider,Nazvanie,Fio,Number,Adress")] Pprovider pprovider)
+        public async Task<IActionResult> Edit(long id, [Bind("IdProvider,Nazvanie,Fio,Number,Adress")] Pprovider pprovider)
         {
             if (id != pprovider.IdProvider)
             {
@@ -118,7 +118,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Pproviders/Delete/5
-        public async Task<IActionResult> Delete(decimal? id)
+        public async Task<IActionResult> Delete(long? id)
         {
             if (id == null || _context.Pproviders == null)
             {
@@ -138,7 +138,7 @@ namespace StoreASP.Controllers
         // POST: Pproviders/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(decimal id)
+        public async Task<IActionResult> DeleteConfirmed(long id)
         {
             if (_context.Pproviders == null)
             {
@@ -154,7 +154,7 @@ namespace StoreASP.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool PproviderExists(decimal id)
+        private bool PproviderExists(long id)
         {
           return (_context.Pproviders?.Any(e => e.IdProvider == id)).GetValueOrDefault();
         }
