+++ StoreASP/Controllers/ProductSalesController.cs
@@ -26,7 +26,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: ProductSales/Details/5
-        public async Task<IActionResult> Details(decimal? id)
+        public async Task<IActionResult> Details(long? id)
         {
             if (id == null || _context.ProductSales == null)
             {
@@ -72,7 +72,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: ProductSales/Edit/5
-        public async Task<IActionResult> Edit(decimal? id)
+        public async Task<IActionResult> Edit(long? id)
         {
             if (id == null || _context.ProductSales == null)
             {
@@ -94,7 +94,7 @@ namespace StoreASP.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(decimal id, [Bind("IdProduct,IdSale,KolVo2")] ProductSale productSale)
+        public async Task<IActionResult> Edit(long id, [Bind("IdProduct,IdSale,KolVo2")] ProductSale productSale)
         {
             if (id != productSale.IdProduct)
             {
@@ -127,7 +127,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: ProductSales/Delete/5
-        public async Task<IActionResult> Delete(decimal? id)
+        public async Task<IActionResult> Delete(long? id)
         {
             if (id == null || _context.ProductSales == null)
             {
@@ -149,7 +149,7 @@ namespace StoreASP.Controllers
         // POST: ProductSales/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(decimal id)
+        public async Task<IActionResult> DeleteConfirmed(long id)
         {
             if (_context.ProductSales == null)
             {
@@ -165,7 +165,7 @@ namespace StoreASP.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool ProductSaleExists(decimal id)
+        private bool ProductSaleExists(long id)
         {
           return (_context.ProductSales?.Any(e => e.IdProduct == id)).GetValueOrDefault();
         }
