+++ StoreASP/Controllers/ProductsController.cs
@@ -26,7 +26,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Products/Details/5
-        public async Task<IActionResult> Details(decimal? id)
+        public async Task<IActionResult> Details(long? id)
         {
             if (id == null || _context.Products == null)
             {
@@ -72,7 +72,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Products/Edit/5
-        public async Task<IActionResult> Edit(decimal? id)
+        public async Task<IActionResult> Edit(long? id)
         {
             if (id == null || _context.Products == null)
             {
@@ -94,7 +94,7 @@ namespace StoreASP.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(decimal id, [Bind("IdProduct,NazvanieProduct,Opisanie,Foto,CostSales,Imei,IdMark,IdCategory")] Product product)
+        public async Task<IActionResult> Edit(long id, [Bind("IdProduct,NazvanieProduct,Opisanie,Foto,CostSales,Imei,IdMark,IdCategory")] Product product)
         {
             if (id != product.IdProduct)
             {
@@ -127,7 +127,7 @@ namespace StoreASP.Controllers
         }
 
         // GET: Products/Delete/5
-        public async Task<IActionResult> Delete(decimal? id)
+        public async Task<IActionResult> Delete(long? id)
         {
             if (id == null || _context.Products == null)
             {
@@ -149,7 +149,7 @@ namespace StoreASP.Controllers
         // POST: Products/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(decimal id)
+        public async Task<IActionResult> DeleteConfirmed(long id)
         {
             if (_context.Products == null)
             {
@@ -165,7 +165,7 @@ namespace StoreASP.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool ProductExists(decimal id)
+        private bool ProductExists(long id)
         {
           return (_context.Products?.Any(e => e.IdProduct == id)).GetValueOrDefault();
         }
